+++ packages/components/src/accordion/AccordionSummary.tsx
@@ -1,15 +1,16 @@
 import * as React from 'react'
 import { useAccordionContext } from './context/AccordionContext'
 import { cns, appliedStyles } from '@stcland/utils'
+import { AccordionStyles } from 'src/styles'
 
 //*****************************************************************************
 // Interface
 //*****************************************************************************
 
 export interface AccordionSummaryProps extends React.HTMLAttributes<HTMLDivElement> {
-  children: React.ReactNode;
-  expandIcon?: React.ReactNode;
-  className?: string;
+  children: React.ReactNode; // Content rendered as the `header` of the accordion section
+  expandIcon?: React.ReactNode; // Icon displayed to indicate `expand/collapse` state
+  className?: string; // Additonal classes for styling applied to the `root`
 }
 
 //*****************************************************************************
@@ -24,12 +25,12 @@ export const AccordionSummary = ({
 }: AccordionSummaryProps) => {
   const { expanded, toggle, summaryId, detailsId, customStyles } = useAccordionContext()
 
-  const accordionSummaryStyles = {
+  const accordionSummaryStyles : AccordionStyles['accordionSummary'] = {
     root: cns('flex justify-between items-center p-4 cursor-pointer bg-gray-825 text-gray-100 border-b border-primary-main', className),
     icon: 'ml-2 transition-transform duration-300 ease-in-out'
   }
 
-  const mergedStyles = appliedStyles(accordionSummaryStyles, customStyles?.accordionSummary)
+  const mergedStyles = appliedStyles<AccordionStyles['accordionSummary']>(accordionSummaryStyles, customStyles?.accordionSummary)
 
   const cn = {
     root: mergedStyles.root,