+++ packages/components/src/styles/index.ts
@@ -16,4 +16,5 @@ export type {
   FormStyles,
   FieldSetStyles,
   FormTitleStyles,
+  ListStyles
 } from './componentTypes'