+++ packages/components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stcland/components
 
+## 0.0.8
+
+### Patch Changes
+
+- Some enhancements in List composition.
+
+## 0.0.7
+
+### Patch Changes
+
+- Added List and Form to STCLandia.
+
 ## 0.0.6
 
 ### Patch Changes